Refactor: refactor ts file name
