+++ rust/scripts/buildAll.ts
