+++ rust/scripts/src/scripts/canisters/dfBurnable.ts
