+++ rust/scripts/src/scripts/canisters/dftBasic.ts
