+++ rust/scripts/src/scripts/canisters/dftBasic2.ts
