+++ rust/scripts/src/scripts/canisters/dftMintable.ts
