+++ rust/scripts/src/scripts/canisters/index.ts
@@ -1,3 +1,3 @@
-import {get_dfx_json} from "~/utils/dfx_json";
+import {get_dfx_json} from "~/utils/dfxJson";
 
 export const canisters = Array.from(get_dfx_json().canisters);
