+++ rust/scripts/updateDid.ts
