+++ rust/scripts/updateLocalConfigs.ts
