+++ rust/sh_build.sh
@@ -1,1 +1,1 @@
-ts-node -r tsconfig-paths/register scripts/build_all.ts
+ts-node -r tsconfig-paths/register scripts/buildAll.ts
