+++ rust/sh_did.sh
@@ -1,1 +1,1 @@
-ts-node -r tsconfig-paths/register scripts/update_did.ts
+ts-node -r tsconfig-paths/register scripts/updateDid.ts
