+++ rust/sh_setup_dev.sh
@@ -4,5 +4,5 @@ npm i ts-node -g
 npm i tsconfig-paths -g
 npm i jake -g
 npm i @cucumber/cucumber -g
-ts-node -r tsconfig-paths/register scripts/create_identities.ts
-ts-node -r tsconfig-paths/register scripts/update_local_configs.ts
+ts-node -r tsconfig-paths/register scripts/createIdentities.ts
+ts-node -r tsconfig-paths/register scripts/updateLocalConfigs.ts
